requirements.txt
biopython==1.78
h5py==2.10
matplotlib==3.1.3
numpy==1.18.1
psutil==5.6.7
pybedtools==0.8.1
pydot==1.4.1
scikit-learn==0.22.1
six==1.14.0
statsmodels==0.11.0
shap==0.35
tensorboard==2.1.0
tensorflow==2.1.0
weblogo==3.7.4
pandas==1.0.3
scipy==1.4.1
seaborn==0.11
tqdm==4.49
